charms/openfga-k8s/requirements.txt
ops >= 1.5.0
ops-lib-pgsql
charmhelpers >= 0.20.22
